+++ jax/lib/xla_client.py
@@ -25,7 +25,6 @@ shape_from_pyval = _xc.shape_from_pyval
 ArrayImpl = _xc.ArrayImpl
 Client = _xc.Client
 CompileOptions = _xc.CompileOptions
-Device = _xc.Device
 DeviceAssignment = _xc.DeviceAssignment
 FftType = _xc.FftType
 Frame = _xc.Frame
@@ -49,6 +48,10 @@ _deprecations = {
         _xc.bfloat16,
     ),
     # Added Sep 26 2024
+    "Device" : (
+      "jax.lib.xla_client.Device is deprecated; use jax.Device instead.",
+      _xc.Device
+    ),
     "XlaRuntimeError": (
         (
             "jax.lib.xla_client.XlaRuntimeError is deprecated; use"
@@ -63,6 +66,7 @@ import typing as _typing
 if _typing.TYPE_CHECKING:
   _xla = _xc._xla
   bfloat16 = _xc.bfloat16
+  Device = _xc.Device
   XlaRuntimeError = _xc.XlaRuntimeError
 else:
   from jax._src.deprecations import deprecation_getattr as _deprecation_getattr